+++ src/main/kotlin/cat/wavy/catactivity/setting/DefaultsState.kt
@@ -3,12 +3,9 @@ package cat.wavy.catactivity.setting
 import cat.wavy.catactivity.types.IDEType
 import cat.wavy.catactivity.types.currentIDEType
 
-data class SettingState(
-    var isEnabled: Boolean = false,
-    var details: Details = Details.IDE,
+data class DefaultsState(
     var theme: Theme = Theme.Macchiato,
     var ideIcon: IDEIcon = IDEIcon.New,
-    var showRepositoryButton: Boolean = false,
 
     var projectStateFormat: String = "Working on %projectName%",
     var projectDetailFormat: String = "%projectProblems% Problems in Project",
@@ -18,15 +15,7 @@ data class SettingState(
 
     var idleStateFormat: String = "Idle",
     var idleDetailFormat: String = "%projectName% (%branch%)",
-
-    var firstInit: Boolean = true,
 )
-
-enum class Details {
-    IDE,
-    Project,
-    File,
-}
 
 enum class Theme {
     Frappe,
@@ -41,3 +30,4 @@ enum class IDEIcon(val displayName: String, val app: IDEType) {
     JetBrains("JetBrains", IDEType.JETBRAINS),
     CatActivity("Cat Activity", IDEType.CATACTIVITY),
 }
+